avances y commit para checkout
