+++ .env
@@ -1,5 +1,5 @@
-NEXT_PUBLIC_API_SENDER_SOCKET_URL="https://test3.dragonchat.io/"
-NEXT_PUBLIC_API_SENDER_URL="https://test3.dragonchat.io/api/v1"
+NEXT_PUBLIC_API_SENDER_SOCKET_URL="https://api-sender.dragonchat.io/"
+NEXT_PUBLIC_API_SENDER_URL="https://api-sender.dragonchat.io/api/v1"
 NEXT_PUBLIC_API_USER_URL="https://api-user.dragonchat.io/api/v1"
 NEXT_PUBLIC_LOGIN_COOKIE_NAME="dragonchat_login"
 PORT=5004
