+++ components/cards/ConversationPremium/ConversationPremium.module.css
@@ -450,5 +450,5 @@
     width: 100%;
     height: 100%;
     z-index: 4;
-    background-color: rgba(0, 0, 0, 0.3);
+    background-color: rgba(0, 0, 0, 0.8);
 }
